pull out conversation formatting to a standalone object
